Added requirement that the beambreak sensor has changed state at least once after the last dispense timeout FOR FOOD PORTS
Similar to the implementation for the water ports, we added the requirement that the food beambreaks change state before another sequential food dispense event will be allowed to occur. Failure to implement this previously resulted in an anomalous number of dispense events without a preceding beambreak event due to debris stuck in the port.